+++ control/SLM/slmpy.py
@@ -88,7 +88,7 @@ class SLMdisplay(object):
 
     def getSize(self):
         return self.vt.frame._resX, self.vt.frame._resY
-
+        
     def updateArray(self, ev_mask):
         """Update the SLM monitor with the supplied EvMask.
         Note that the array is not the same size as the SLM resolution,
